client and server fixed for work with windows

+++ server.py
@@ -9,7 +9,8 @@ port = int(sys.argv[1])
 
 clients_id_path = {}  # create dictionary that maps id to path in the server.
 data_base = {}
-
+server_has_changed = True  # global var that check if the server has been changed from last push request.
+my_ops = os.name
 # open server
 main_socket = socket()
 main_socket.bind(('', port))
@@ -26,6 +27,13 @@ def get_path(src_platform, src_path, src_sep='/'):
     return src_path
 
 
+# def get_path(src_platform, src_path):
+#     if src_path.__contains__('\\') and my_ops == "posix":
+#         src_path = src_path.replace('\\', '/')
+#     elif src_path.__contains__('/') and my_ops == "nt":
+#         src_path = src_path.replace('/', '\\')
+#         return src_path
+
 # TODO change id length to 128 characters
 # return random string with 128 characters compared of letters and numbers.
 def get_random_id(id_set):
@@ -35,6 +43,7 @@ def get_random_id(id_set):
     return new_id
 
 
+# This function get from a new client without id the dirs and files in client path
 def get_files(get_files_sock):
     line = " "
     while True:
@@ -48,12 +57,14 @@ def get_files(get_files_sock):
                 if not line:
                     break
                 dir_name = line.strip().decode()
+                dir_name = get_path(client_op, dir_name)
                 print(f'empty dir {dir_name} ... \n', end='', flush=True)
                 dir_path = os.path.join('AllClients', str(client_id))
                 dir_path = os.path.join(dir_path, dir_name)
                 os.makedirs(dir_path, exist_ok=True)
         else:
             filename = line.strip().decode()
+            filename = get_path(client_op, filename)
             length = int(get_files_sock.readline())
             print(f'Downloading {filename}...\n  Expecting {length:,} bytes...', end='', flush=True)
 
@@ -80,6 +91,7 @@ def get_files(get_files_sock):
     get_files_sock.close()
 
 
+# This function send to new client with id the dirs and files in client path
 def send_files(on_sock, src_path):  # type - socket #send file and empty dirs
     with on_sock:
         for root, dirs, files in os.walk(src_path):
@@ -123,7 +135,7 @@ def create_dirs(d_path):
 def get_file(on_socket, file_path):  # type - makefile('rb')
     file_name = on_socket.readline()
     length = int(on_socket.readline())
-    create_dirs(os.path.dirname(file_path))
+    os.makedirs(os.path.dirname(file_path), exist_ok=True)
     with open(file_path, 'wb') as f:
         while length:
             chunk = min(length, CHUNKSIZE)
@@ -156,23 +168,30 @@ def send_file(on_sock, src_path):
                 on_sock.sendall(data)
 
 
+# give the user the computer number of this user
 def get_comp_num(c_id):
     if data_base.keys().__contains__(c_id):  # check if the ID exists
         c_comp = str(len(data_base[c_id]) + 1)
-        data_base[c_id][c_comp] = []  # new computer has joined the data base.
+        data_base[c_id][c_comp] = list()  # new computer has joined the data base.
     else:
         c_comp = "1"
         data_base[c_id] = {}
-        data_base[c_id][c_comp] = []
+        data_base[c_id][c_comp] = list()
     return c_comp
 
 
 def get_update(command, get_data_sock):
+    global server_has_changed
     if command[0] == "created":
         is_dir, src_path = command[1], command[2]
         src_path = os.path.join(clients_id_path[client_id], src_path)
-        if is_dir == "True" and not os.path.exists(src_path):
-            os.mkdir(src_path)
+        src_path = get_path(client_op, src_path)
+        if os.path.exists(src_path):
+            server_has_changed = False
+            return
+        server_has_changed = True
+        if is_dir == "True":
+            os.makedirs(src_path)
             print("created dir")
         elif is_dir == "False":
             get_file(get_data_sock, src_path)
@@ -180,30 +199,39 @@ def get_update(command, get_data_sock):
     elif command[0] == "deleted":
         is_dir, del_path = command[1], command[2]
         del_path = os.path.join(clients_id_path[client_id], del_path)
+        if not os.path.exists(del_path):
+            server_has_changed = False
+            return
+        server_has_changed = True
         if is_dir == "True":
-            if os.listdir(del_path):
-                print("error")
-            os.rmdir(del_path)
+            delete_dir(del_path)
         else:
-            os.remove(del_path)
+            if os.path.exists(del_path):
+                os.remove(del_path)
 
     elif command[0] == "moved":
         is_dir, src_path, dest_path = command[1], command[2], command[3]
+        src_path = get_path(client_op, src_path)
+        dest_path = get_path(client_op, dest_path)
         src_path = os.path.join(clients_id_path[client_id], src_path)
         dest_path = os.path.join(clients_id_path[client_id], dest_path)
+        if not os.path.exists(src_path) and os.path.exists(dest_path):
+            server_has_changed = False
+            return
+        server_has_changed = True
         if is_dir == "False":
             os.makedirs(os.path.dirname(dest_path), exist_ok=True)
             os.replace(src_path, dest_path)
         else:
             delete_dir(src_path)
             if not os.path.exists(dest_path):
-                os.mkdir(dest_path)
+                os.makedirs(dest_path)
 
 
 def update_computers(c_id, c_comp, cmd):
     for k in data_base[c_id].keys():
         if k != c_comp:
-            list((data_base[c_id][k])).append(str(cmd))
+            data_base[c_id][k].append(cmd)
 
 
 def delete_dir(path_to_del):
@@ -224,19 +252,31 @@ def send_update(cmd, c_path, on_sock):
     if cmd_type == "created":
         notify_created(cmd, c_path, on_sock)
     elif cmd_type == "deleted":
-        pass
+        notify_deleted(cmd, on_sock)
     elif cmd_type == "moved":
-        pass
+        notify_moved(cmd, on_sock)
 
 
 def notify_created(curr_update, c_path, on_sock):
-    mode, is_dir, new_path = curr_update
+    mode, is_dir, new_path = curr_update.split(',')
     print(curr_update)
     with on_sock:
         on_sock.sendall(curr_update.encode() + b'\n')
         if is_dir == "False":
             new_path = os.path.join(c_path, new_path)
             send_file(on_sock, new_path)
+
+
+def notify_deleted(curr_update, on_sock):
+    print(curr_update)
+    with on_sock:
+        on_sock.sendall(curr_update.encode() + b'\n')
+
+
+def notify_moved(curr_update, on_sock):
+    print(curr_update)
+    with on_sock:
+        on_sock.sendall(curr_update.encode() + b'\n')
 
 
 if __name__ == "__main__":
@@ -285,11 +325,13 @@ if __name__ == "__main__":
             else:
                 connection_type = get_data_sock.readline().strip().decode()
                 if connection_type == "pull":
-                    # already signed computer- check if there are updates:
+                    # already signed computer - check if there are updates:
                     if data_base[client_id][client_comp]:  # check if there are updates
-                        status = str(len(data_base[client_id][client_comp])) + "To go"
+                        status = str(len(data_base[client_id][client_comp])) + " To go!"
                         client_socket.sendall(status.encode() + b'\n')
-                        update = data_base[client_id][client_comp].pop
+                        update = data_base[client_id][client_comp].pop(0)
+                        print(status)
+                        print(update)
                         send_update(update, client_dir_path, client_socket)
                         # TODO: share the updates to this computer
                     else:
@@ -300,8 +342,9 @@ if __name__ == "__main__":
                     command_txt = get_data_sock.readline().strip().decode()
                     command = str(command_txt).split(',')
                     get_update(command, get_data_sock)
-                    update_computers(client_id, client_comp, command_txt)
-
-
-
-
+                    if server_has_changed:
+                        command[2] = get_path(client_op, command[2])
+                        if command[0] == "moved":
+                            command[3] = get_path(client_op, command[3])
+                        curr_command = ','.join(command)
+                        update_computers(client_id, client_comp, curr_command)
